dix dev fixtures and move sites ModelAdmin regsitration to sites.admin
